Extract to python package 🧨
